requirements.txt
flask
redis
mysql-connector-python




